requirements.txt
flask
inject
flask-lambda
PyJWT